v1 - baloo handler is working. need to fix tests
